Created InputManager to manage Controller objects
All Controller objects in the game that are responsible for handling
various gamepad inputs are now located in the Input Manager. This class
has the purpose of initializing Controller objects based on the type of
gamepad plugged in, as well as updating them. This abstraction now
allows us to keep track of the last updated/currently active controller,
which will be useful when programming controls for the various menus in
the game.

Oh, the Item Manager is now called the "Data Manager" as I might need to
place other non-item related information there in the future.